requirements.txt
asgiref==3.7.2
certifi==2023.5.7
Django==4.1.9
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
pymongo==3.12.3
pytz==2023.3
sqlparse==0.2.4
tzdata==2023.3
